+++ nfse-danfse/src/main/java/br/com/nfse/danfse/Danfse.java
@@ -11,6 +11,7 @@ import java.time.OffsetDateTime;
  */
 public record Danfse(
     String chaveAcesso,
+    boolean homologacao,
     Identificacao identificacao,
     Pessoa prestador,
     Pessoa tomador,